Move SearchRequestParser into pkg/api/search.go
